refactor(tests/update): one integration scenario per file
Split update_check.py (2 tests), update_help.py (3 tests),
update_setcap_message.py (2 tests), and update_patterns.py (4 tests)
into individual files — one test function each — matching the convention
used across the rest of the tests/ tree.

No test logic changed; only file organisation.
